configurando controllador de categorias

+++ database/migrations/2024_11_26_220935_create_movements_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('producto_id')->constrained('productos')->onDelete('cascade');
             $table->enum('tipo', ['entrada', 'salida']);
-            $table->float('cantidad', 13, 3);
+            $table->integer('cantidad');
             $table->foreignId('user_id')->nullable()->constrained('users')->onDelete('set null');
             $table->timestamps();
         });
